refactor: review fixes, playback perf & prefetch-by-need
Review findings (profile repo / jni / logging):
- likedTracks: drop duplicate LIKE_LIST round-trip
- ApiResult.code Double -> Int (jni ctor (II[B)V, comparisons)
- urlCache: 6h TTL instead of unbounded; fail-slow cached too
- PlayerHolder recoveryScope lazy init; ProfileViewModel await once
- NumeApp: clear lifted list-island state on leaving detail (DisposableEffect)
- WebLoginScreen vm param no default; ProfileScreen indent
- gate ProfileDiag debug logs behind BuildConfig.DEBUG
- Track gains albumName; shared TrackParser (Track/parseTrack/parseTracks),
  Chart/Profile dedupe; parsePlaylistObject/parseAlbumObject drop fn param

Playback performance:
- rememberPlayerState: split positionMs into its own snapshot; poll only
  writes position, meta updated on real change; NumeApp uses lightweight
  rememberHasTrack so the nav host no longer recomposes at 250ms
- AsyncImage explicit target sizes (track row 96, header 480, playlist 320)
  + remember(Uri); TrackRow base modifier as top-level constant

Request pressure:
- enqueueRest -> prefetch-by-need: keep only current + 5 ready items, poll
  500ms, requests trickle with playback instead of parsing the whole list
  at once (JNI crypto is CPU-heavy; 115-track play went 115->6 requests)

+++ app/src/main/java/com/thripleq/nume/core/net/ApiResult.kt
@@ -7,7 +7,7 @@ package com.thripleq.nume.core.net
  * body has no `code` field.
  */
 class ApiResult(
-    @JvmField val code: Double,
+    @JvmField val code: Int,
     @JvmField val err: Int,
     @JvmField val body: ByteArray,
 )
